Added Book,Liberian,Member and Loan flows

+++ src/main/java/com/aysavs/libraryApp/domain/aggragate/liberian/Liberian.java
@@ -16,7 +16,7 @@ public class Liberian {
     @Column(name="id")  @GeneratedValue(strategy = GenerationType.IDENTITY)private long id;
     @Column(name="name")private String name;
     @Column(name="surname")private String surname;
-    @Column(name="gsmNumber")private int gsmNumber;
+    @Column(name="gsmNumber")private long gsmNumber;
     @Column(name="mailAddress")private String mailAddress;
